Made SnippetPreviewSection a presentation component only and moved SnippetEditor container logic to the the SnippetEditor container
